mitosheet: update test to create valid dataframe

+++ mitosheet/mitosheet/tests/steps_performers/test_filter.py
@@ -51,7 +51,7 @@ FILTER_TESTS = [
         NUMBER_SERIES,
         FC_NUMBER_EXACTLY,
         10,
-        pd.DataFrame(columns=['A']),
+        pd.DataFrame(columns=['A'], dtype='int'),
     ),
     (
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
@@ -65,7 +65,7 @@ FILTER_TESTS = [
         NUMBER_SERIES,
         FC_NUMBER_GREATER,
         10,
-        pd.DataFrame(columns=['A']),
+        pd.DataFrame(columns=['A'], dtype='int'),
     ), 
     (
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
@@ -86,7 +86,7 @@ FILTER_TESTS = [
         NUMBER_SERIES,
         FC_NUMBER_GREATER_THAN_OR_EQUAL,
         10,
-        pd.DataFrame(columns=['A']),
+        pd.DataFrame(columns=['A'], dtype='int'),
     ), 
     (
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
@@ -107,7 +107,7 @@ FILTER_TESTS = [
         NUMBER_SERIES,
         FC_NUMBER_LESS,
         0,
-        pd.DataFrame(columns=['A']),
+        pd.DataFrame(columns=['A'], dtype='int'),
     ), 
     (
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
@@ -128,7 +128,7 @@ FILTER_TESTS = [
         NUMBER_SERIES,
         FC_NUMBER_LESS_THAN_OR_EQUAL,
         0,
-        pd.DataFrame(columns=['A']),
+        pd.DataFrame(columns=['A'], dtype='int'),
     ), 
     (
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
